Managing Ride lifecycle (#27)
add: service-communication

+++ promocode-service/src/main/java/com/dulfinne/taxi/promocodeservice/model/PromocodeUsage.java
@@ -26,5 +26,5 @@ public class PromocodeUsage {
   private String code;
   private String username;
   private Instant usageDate;
-  private Integer rideId;
+  private Long rideId;
 }
